fix(search-filter): code move in new component
