+++ src/components/App/SideBar/index.tsx
@@ -1,14 +1,13 @@
-import { Button, Slide } from '@mui/material'
-import Popover from '@mui/material/Popover'
+import { Slide } from '@mui/material'
 import clsx from 'clsx'
 import React, { forwardRef, useEffect, useRef, useState } from 'react'
 import { useFormContext } from 'react-hook-form'
 import { ClipLoader } from 'react-spinners'
 import styled from 'styled-components'
 import { SelectWithPopover } from '~/components/App/SideBar/Dropdown'
+import { FilterSearch } from '~/components/App/SideBar/FilterSearch'
 import ChevronLeftIcon from '~/components/Icons/ChevronLeftIcon'
 import ClearIcon from '~/components/Icons/ClearIcon'
-import PlusIcon from '~/components/Icons/PlusIcon'
 import SearchFilterCloseIcon from '~/components/Icons/SearchFilterCloseIcon'
 import SearchFilterIcon from '~/components/Icons/SearchFilterIcon'
 import SearchIcon from '~/components/Icons/SearchIcon'
@@ -16,7 +15,6 @@ import { SchemaExtended } from '~/components/ModalsContainer/BlueprintModal/type
 import { SearchBar } from '~/components/SearchBar'
 import { Flex } from '~/components/common/Flex'
 import { FetchLoaderText } from '~/components/common/Loader'
-import { api } from '~/network/api'
 import { getSchemaAll } from '~/network/fetchSourcesData'
 import { useAppStore } from '~/stores/useAppStore'
 import { useDataStore, useFilteredNodes, useSelectedNode } from '~/stores/useDataStore'
@@ -54,12 +52,8 @@ const Content = forwardRef<HTMLDivElement, ContentProp>(({ onSubmit, subViewOpen
   const [isFilterOpen, setIsFilterOpen] = useState(false)
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null)
   const [schemaAll, setSchemaAll] = useState<SchemaExtended[]>([])
-  const [selectedTypes, setSelectedTypes] = useState<string[]>([]) // State to store selected node types
+  const [selectedTypes, setSelectedTypes] = useState<string[]>([])
   const [showAllSchemas, setShowAllSchemas] = useState(false)
-
-  const handleViewMoreClick = () => {
-    setShowAllSchemas(true)
-  }
 
   useEffect(() => {
     setValue('search', searchFormValue)
@@ -81,12 +75,8 @@ const Content = forwardRef<HTMLDivElement, ContentProp>(({ onSubmit, subViewOpen
 
   const typing = watch('search')
 
-  const handleFilterIconClick = async (event: React.MouseEvent<HTMLElement>) => {
-    if (isFilterOpen) {
-      setAnchorEl(null)
-    } else {
-      setAnchorEl(event.currentTarget)
-
+  useEffect(() => {
+    const fetchSchemaData = async () => {
       try {
         const response = await getSchemaAll()
 
@@ -96,33 +86,19 @@ const Content = forwardRef<HTMLDivElement, ContentProp>(({ onSubmit, subViewOpen
       }
     }
 
+    fetchSchemaData()
+  }, [])
+
+  const handleFilterIconClick = (event: React.MouseEvent<HTMLElement>) => {
+    if (isFilterOpen) {
+      setAnchorEl(null)
+    } else {
+      setAnchorEl(event.currentTarget)
+    }
+
     setIsFilterOpen((prev) => !prev)
     setSelectedTypes([])
     setShowAllSchemas(false)
-  }
-
-  const handleSchemaTypeClick = (type: string) => {
-    setSelectedTypes((prevSelectedTypes) =>
-      prevSelectedTypes.includes(type) ? prevSelectedTypes.filter((t) => t !== type) : [...prevSelectedTypes, type],
-    )
-  }
-
-  const handleClear = async () => {
-    setSelectedTypes([])
-  }
-
-  const handleSave = async () => {
-    const endPoint = 'graph/search'
-
-    const queryParams = new URLSearchParams({
-      node_type: selectedTypes.join(','),
-    })
-
-    try {
-      await api.get(`/${endPoint}?${queryParams.toString()}`)
-    } catch (error) {
-      console.warn(error)
-    }
   }
 
   return (
@@ -162,89 +138,16 @@ const Content = forwardRef<HTMLDivElement, ContentProp>(({ onSubmit, subViewOpen
           <IconWrapper isFilterOpen={isFilterOpen} onClick={handleFilterIconClick}>
             {isFilterOpen ? <SearchFilterCloseIcon /> : <SearchFilterIcon />}
           </IconWrapper>
-          <SearchFilterPopover
+          <FilterSearch
             anchorEl={anchorEl}
-            anchorOrigin={{
-              vertical: 'bottom',
-              horizontal: 'left',
-            }}
-            anchorPosition={{
-              top: 200,
-              left: 0,
-            }}
-            onClose={() => {
-              setAnchorEl(null)
-              setIsFilterOpen(false)
-            }}
-            open={Boolean(anchorEl)}
-            PaperProps={{
-              style: {
-                marginTop: '13px',
-                marginLeft: '30px',
-              },
-            }}
-            transformOrigin={{
-              vertical: 'top',
-              horizontal: 'right',
-            }}
-          >
-            <PopoverHeader>
-              <div>Type</div>
-              <div>{selectedTypes.length} Selected</div>
-            </PopoverHeader>
-            <PopoverBody>
-              <SchemaTypeWrapper>
-                {(showAllSchemas ? schemaAll : schemaAll.slice(0, 15)).map((schema) => (
-                  <SchemaType
-                    key={schema.type}
-                    isSelected={selectedTypes.includes(schema.type as string)}
-                    onClick={() => handleSchemaTypeClick(schema?.type as string)}
-                  >
-                    {schema.type}
-                  </SchemaType>
-                ))}
-              </SchemaTypeWrapper>
-              {!showAllSchemas && schemaAll.length > 15 && (
-                <ViewMoreButton onClick={handleViewMoreClick}>
-                  <PlusIconWrapper>
-                    <PlusIcon /> View More
-                  </PlusIconWrapper>
-                </ViewMoreButton>
-              )}
-            </PopoverBody>
-            <LineBar />
-            <PopoverFooter>
-              <Flex direction="row" mb={6}>
-                <DeleteButton
-                  color="secondary"
-                  // disabled={topicIsLoading || !node}
-                  onClick={handleClear}
-                  size="large"
-                  style={{ marginRight: 20 }}
-                  variant="contained"
-                >
-                  <ClearButtonWrapper>
-                    <ClearIcon />
-                  </ClearButtonWrapper>
-                  Clear
-                </DeleteButton>
-                <ShowResultButton
-                  color="secondary"
-                  // disabled={shouldDisableSave}
-                  onClick={handleSave}
-                  size="large"
-                  variant="contained"
-                >
-                  Show Results
-                  {/* {loading && (*/}
-                  {/*    <ClipLoaderWrapper>*/}
-                  {/*      <ClipLoader color={colors.lightGray} size={12} />*/}
-                  {/*    </ClipLoaderWrapper>*/}
-                  {/* )}*/}
-                </ShowResultButton>
-              </Flex>
-            </PopoverFooter>
-          </SearchFilterPopover>
+            schemaAll={schemaAll}
+            selectedTypes={selectedTypes}
+            setAnchorEl={setAnchorEl}
+            setIsFilterOpen={setIsFilterOpen}
+            setSelectedTypes={setSelectedTypes}
+            setShowAllSchemas={setShowAllSchemas}
+            showAllSchemas={showAllSchemas}
+          />
         </SearchFilterIconWrapper>
         {searchTerm && (
           <SearchDetails>
@@ -464,146 +367,6 @@ const IconWrapper = styled.div<{ isFilterOpen: boolean }>`
     height: ${({ isFilterOpen }) => (isFilterOpen ? '11px' : '24px')};
     color: ${({ isFilterOpen }) => (isFilterOpen ? colors.black : colors.GRAY7)};
     fill: none;
-  }
-`
-
-const SearchFilterPopover = styled(Popover)`
-  .MuiPaper-root {
-    background: ${colors.BG2};
-    padding: 16px;
-    min-width: 360px;
-    color: ${colors.white};
-    box-shadow: 0px 1px 6px 0px rgba(0, 0, 0, 0.2);
-    border-radius: 9px;
-    max-width: 361px;
-    border: 2px solid ${colors.black};
-  }
-`
-
-const PopoverHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-bottom: 8px;
-  font-weight: bold;
-`
-
-const PopoverBody = styled.div`
-  padding: 16px 0;
-`
-
-const PopoverFooter = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-top: 8px;
-`
-
-const LineBar = styled.div`
-  border: 1px solid ${colors.black};
-  margin: 4px auto;
-`
-
-const PlusIconWrapper = styled.span`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 6px;
-`
-
-const SchemaTypeWrapper = styled(Flex).attrs({
-  align: 'center',
-  direction: 'row',
-  grow: 1,
-  justify: 'flex-start',
-})`
-  flex-wrap: wrap;
-  gap: 10px;
-  max-height: 400px;
-  overflow: auto;
-`
-
-const SchemaType = styled(Flex).attrs({
-  align: 'center',
-  direction: 'row',
-  justify: 'flex-start',
-})<{ isSelected: boolean }>`
-  color: ${({ isSelected }) => (isSelected ? colors.black : colors.white)};
-  background: ${({ isSelected }) => (isSelected ? colors.white : colors.BUTTON1_PRESS)};
-  padding: 6px 10px 6px 8px;
-  font-family: Barlow;
-  font-size: 13px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 15px;
-  letter-spacing: 0.78px;
-  margin: 0 8px;
-  border-radius: 200px;
-  cursor: pointer;
-
-  &:hover {
-    background: ${({ isSelected }) => (isSelected ? colors.white : colors.BUTTON1_PRESS)};
-  }
-
-  &:active {
-    background: ${colors.white};
-    color: ${colors.black};
-  }
-`
-
-const ViewMoreButton = styled.button`
-  background: transparent;
-  color: ${colors.white};
-  border: none;
-  padding: 8px 12px;
-  margin-top: 10px;
-  cursor: pointer;
-  border-radius: 4px;
-  font-family: Barlow;
-  font-size: 13px;
-  font-weight: 500;
-
-  &:hover {
-    background: ${colors.BUTTON1_HOVER};
-  }
-
-  &:active {
-    background: ${colors.BUTTON1_PRESS};
-  }
-`
-
-// const ClipLoaderWrapper = styled.span`
-//   margin-top: 3px;
-// `
-
-const DeleteButton = styled(Button)`
-  && {
-    color: ${colors.white};
-    background-color: ${colors.BUTTON1};
-    padding-left: 4px;
-    &:hover,
-    &:active,
-    &:focus {
-      color: rgba(255, 255, 255, 0.85);
-      background-color: ${colors.BUTTON1};
-    }
-  }
-`
-
-const ClearButtonWrapper = styled.span`
-  svg {
-    width: 32px;
-    height: 32px;
-    color: ${colors.GRAY7};
-    fill: none;
-    margin-top: 2px;
-  }
-`
-
-const ShowResultButton = styled(Button)`
-  && {
-    flex: 1;
-    padding: 2px 55px;
   }
 `
